Resolve reoccurrence of issue #78 for admin list display converts
